Moved Helper Classes from TestImplementations.java to separate Files
