+++ src/main/java/org/jqassistant/contrib/plugin/csharp/json_to_neo4j/ParameterAnalyzer.java
@@ -51,6 +51,4 @@ public class ParameterAnalyzer {
         parameterDescriptor.setType(parameterTypeDescriptor);
         parameterDescriptor.setName(parameterModel.getName());
     }
-
-
 }
