+++ src/test/java/org/jqassistant/contrib/plugin/csharp/json_to_neo4j/testImplementations/ConstructorDescriptorImpl.java
@@ -1,0 +1,6 @@
+package org.jqassistant.contrib.plugin.csharp.json_to_neo4j.testImplementations;
+
+import org.jqassistant.contrib.plugin.csharp.model.ConstructorDescriptor;
+
+public class ConstructorDescriptorImpl extends MethodDescriptorImpl implements ConstructorDescriptor {
+}
