+++ src/test/java/org/jqassistant/contrib/plugin/csharp/json_to_neo4j/testImplementations/EnumDescriptorImpl.java
@@ -1,0 +1,9 @@
+package org.jqassistant.contrib.plugin.csharp.json_to_neo4j.testImplementations;
+
+import org.jqassistant.contrib.plugin.csharp.model.EnumTypeDescriptor;
+
+public class EnumDescriptorImpl extends ClassDescriptorImpl implements EnumTypeDescriptor {
+    public EnumDescriptorImpl(String name) {
+        super(name);
+    }
+}
