+++ src/test/java/org/jqassistant/contrib/plugin/csharp/json_to_neo4j/testImplementations/ParameterDescriptorImpl.java
@@ -1,0 +1,57 @@
+package org.jqassistant.contrib.plugin.csharp.json_to_neo4j.testImplementations;
+
+import org.jqassistant.contrib.plugin.csharp.model.ParameterDescriptor;
+import org.jqassistant.contrib.plugin.csharp.model.TypeDescriptor;
+
+public class ParameterDescriptorImpl implements ParameterDescriptor {
+
+    private int index;
+    private String name;
+    private TypeDescriptor type;
+
+    @Override
+    public TypeDescriptor getType() {
+        return type;
+    }
+
+    @Override
+    public void setType(TypeDescriptor type) {
+        this.type = type;
+    }
+
+    @Override
+    public int getIndex() {
+        return index;
+    }
+
+    @Override
+    public void setIndex(int index) {
+        this.index = index;
+    }
+
+    @Override
+    public String getName() {
+        return name;
+    }
+
+    @Override
+    public void setName(String name) {
+        this.name = name;
+    }
+
+
+    @Override
+    public <I> I getId() {
+        return null;
+    }
+
+    @Override
+    public <T> T as(Class<T> aClass) {
+        return null;
+    }
+
+    @Override
+    public <D> D getDelegate() {
+        return null;
+    }
+}
