+++ src/test/java/org/jqassistant/contrib/plugin/csharp/json_to_neo4j/testImplementations/PrimitiveValueDescriptorImpl.java
@@ -1,0 +1,56 @@
+package org.jqassistant.contrib.plugin.csharp.json_to_neo4j.testImplementations;
+
+import org.jqassistant.contrib.plugin.csharp.model.PrimitiveValueDescriptor;
+import org.jqassistant.contrib.plugin.csharp.model.TypeDescriptor;
+
+public class PrimitiveValueDescriptorImpl implements PrimitiveValueDescriptor {
+
+    private String name;
+    private TypeDescriptor type;
+    private Object value;
+
+    @Override
+    public String getName() {
+        return name;
+    }
+
+    @Override
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    @Override
+    public TypeDescriptor getType() {
+        return type;
+    }
+
+    @Override
+    public void setType(TypeDescriptor type) {
+        this.type = type;
+    }
+
+    @Override
+    public Object getValue() {
+        return value;
+    }
+
+    @Override
+    public void setValue(Object value) {
+        this.value = value;
+    }
+
+    @Override
+    public <I> I getId() {
+        return null;
+    }
+
+    @Override
+    public <T> T as(Class<T> aClass) {
+        return null;
+    }
+
+    @Override
+    public <D> D getDelegate() {
+        return null;
+    }
+}
